requirements.txt
scikit-learn
pandas
python-dotenv
pymongo
matplotlib
certifi
pymongo[srv]==3.12
#-e.